aws-bedrock-chat-agent-mcp/agentcore/src/requirements.txt
# AgentCore MCP Server - Unified Requirements
# Supports AWS (Bedrock, S3, DynamoDB) and SharePoint integrations

# Web Framework & Server
uvicorn[standard]

# AWS SDK
boto3>=1.35.0
botocore>=1.35.0

# MCP Framework
fastmcp

# AWS Bedrock AgentCore Identity (for OAuth2 Credential Provider)
bedrock-agentcore

# HTTP Client
requests>=2.31.0

# Data Validation
pydantic>=2.0.0

# Document Processing (for SharePoint)
python-docx>=0.8.11
PyPDF2>=3.0.0
openpyxl>=3.1.0
pandas>=1.5.0

# Logging & Configuration
python-json-logger
python-dotenv>=1.0.0